fix: Client got stuck in lobby
Sometimes the client got stuck in the lobby after the game started. If
it was the first player in the first turn then all clients had to quit
and the server had to be restarted. This was probably a race between
starting the game and sending the game update to the clients.