fix: Client sometimes get stuck in lobby
There was a previous fix attempt for this. The client still got stuck in
the lobby sometimes after the game has started. This was due to a race
between the state update handler and the login form handler. These both
relied on the gameStarted flag which has now been removed. Only the
state update handler can switch pages based on the update contents.